minor fixes
fixed some styling issues with coloring of buttons and the like. Also fixed some buttons so that they were properly displayed in forms.

to do - there are some issues with padding and spacing with some of the forms now that i switched over to bootstrap 5. Need to figure out how to pad a button in a wtf quickform etc.